requirements.txt
streamlit
requests
python-dotenv
fpdf 
matplotlib
pandas
